Added project option for init subcommand

+++ cli/src/main/java/com/adaptivescale/rosetta/cli/Cli.java
@@ -45,7 +45,7 @@ class Cli implements Callable<Void> {
         throw new CommandLine.ParameterException(spec.commandLine(), "Missing required subcommand");
     }
 
-    @CommandLine.Command(name = "extract", description = "Extract schema chosen from connection config.")
+    @CommandLine.Command(name = "extract", description = "Extract schema chosen from connection config.", mixinStandardHelpOptions = true)
     private void extract(@CommandLine.Option(names = {"-s", "--source"}, required = true) String sourceName,
                          @CommandLine.Option(names = {"-t", "--convert-to"}) Optional<String> targetName,
                          @CommandLine.Option(names = {"-o", "--output-dir"},
@@ -84,7 +84,7 @@ class Cli implements Callable<Void> {
         log.info("Successfully written output database yaml ({}).", yamlOutputModel.getFilePath());
     }
 
-    @CommandLine.Command(name = "compile", description = "Generate DDL for target Database [bigquery, snowflake, …]")
+    @CommandLine.Command(name = "compile", description = "Generate DDL for target Database [bigquery, snowflake, …]", mixinStandardHelpOptions = true)
     private void compile(@CommandLine.Option(names = {"-t", "--target"}, required = true) String targetName,
                          @CommandLine.Option(names = {"-i", "--input-dir"}, defaultValue = "./") Optional<Path> inputDirectory
     ) throws Exception {
@@ -121,12 +121,25 @@ class Cli implements Callable<Void> {
         log.info("Successfully written ddl ({}).", stringOutput.getFilePath());
     }
 
-    @CommandLine.Command(name = "init", description = "Creates a sample config (main.conf).")
-    private void init() throws IOException {
-        Path fileName = Paths.get(CONFIG_NAME);
+    @CommandLine.Command(name = "init", description = "Creates a sample config (main.conf) and model directory.", mixinStandardHelpOptions = true)
+    private void init(@CommandLine.Parameters(index = "0", description = "Project name.", defaultValue = "")
+                          String projectName) throws IOException {
+        Path fileName = Paths.get(projectName, CONFIG_NAME);
+        Path modelDirectory = Paths.get(projectName, MODEL_DIRECTORY_NAME);
         InputStream resourceAsStream = getClass().getResourceAsStream("/" + TEMPLATE_CONFIG_NAME);
+        Path projectDirectory = Path.of(projectName);
+        if (!projectName.isEmpty() && Files.isDirectory(projectDirectory)) {
+            throw new RuntimeException(String.format("Project (%s) already exists.", projectName));
+        }
+        if (Files.exists(fileName)) {
+            throw new RuntimeException("A configuration for this directory already exists.");
+        }
+        Files.createDirectories(modelDirectory);
         Files.copy(resourceAsStream, fileName);
-        log.info("Successfully created sample config ({}).", CONFIG_NAME);
+        log.info("Successfully created project with a sample config ({}) and model directory ({}).", fileName, modelDirectory);
+        if (!projectName.isEmpty()) {
+            log.info("In order to start using the newly created project please change your working directory.");
+        }
     }
 
     private void requireConfig(Config config) {
